Change Indentation & Reformat, Fix JUnit Error & Continue renaming for consistency
