+++ library/projects/test/build.gradle
@@ -45,6 +45,7 @@ dependencies {
     implementation project(':extract')
     implementation project(':extract-lite')
     implementation 'com.android.support:support-annotations:28.0.0'
+    testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.getkeepsafe.relinker:relinker:1.3.1'
     androidTestImplementation 'com.github.seven332.okio:okio:1.16.0'
