+++ library/src/androidTest/cpp/JavaInit.cpp
@@ -18,52 +18,50 @@
 
 #include "gtest/gtest.h"
 
-TEST(Test, Pass)
-{
-  EXPECT_EQ(1, 1);
+TEST(Test, Pass) {
+    EXPECT_EQ(1, 1);
 }
 
-TEST(Test, Fail)
-{
-  EXPECT_EQ(1, 2);
+TEST(Test, Fail) {
+    EXPECT_EQ(1, 2);
 }
 
 jint a7zip_native_test(
-    JNIEnv* env,
-    jclass,
-    jstring log_file
+        JNIEnv *env,
+        jclass,
+        jstring log_file
 ) {
-  const char* c_log_file = env->GetStringUTFChars(log_file, nullptr);
+    const char *c_log_file = env->GetStringUTFChars(log_file, nullptr);
 
-  std::string xml_path("--gtest_output=xml:");
+    std::string xml_path("--gtest_output=xml:");
 
-  int argc = 2;
-  const char* argv[3] = {"/a7zip_test", (xml_path + c_log_file).c_str(), nullptr};
+    int argc = 2;
+    const char *argv[3] = {"/a7zip_test", (xml_path + c_log_file).c_str(), nullptr};
 
-  testing::InitGoogleTest(&argc, const_cast<char**>(argv));
-  int result = RUN_ALL_TESTS();
+    testing::InitGoogleTest(&argc, const_cast<char **>(argv));
+    int result = RUN_ALL_TESTS();
 
-  env->ReleaseStringUTFChars(log_file, c_log_file);
+    env->ReleaseStringUTFChars(log_file, c_log_file);
 
-  return result;
+    return result;
 }
 
 static JNINativeMethod test_methods[] = {
-    { "nativeTest",
-      "(Ljava/lang/String;)I",
-      (void*) a7zip_native_test }
+        {"nativeTest",
+                "(Ljava/lang/String;)I",
+                (void *) a7zip_native_test}
 };
 
-jint JNI_OnLoad(JavaVM* vm, void*) {
-  JNIEnv* env;
+jint JNI_OnLoad(JavaVM *vm, void *) {
+    JNIEnv *env;
 
-  if (vm->GetEnv(reinterpret_cast<void**>(&env), JNI_VERSION_1_6) != JNI_OK) {
-    return JNI_ERR;
-  }
+    if (vm->GetEnv(reinterpret_cast<void **>(&env), JNI_VERSION_1_6) != JNI_OK) {
+        return JNI_ERR;
+    }
 
-  jclass clazz = env->FindClass("com/hippo/a7zip/NativeTest");
+    jclass clazz = env->FindClass("com/hippo/a7zip/NativeTest");
 
-  env->RegisterNatives(clazz, test_methods, std::extent<decltype(test_methods)>::value);
+    env->RegisterNatives(clazz, test_methods, std::extent<decltype(test_methods)>::value);
 
-  return JNI_VERSION_1_6;
+    return JNI_VERSION_1_6;
 }
