+++ library/src/main/java/com/hippo/a7zip/ArchiveException.java
@@ -21,11 +21,11 @@ import android.support.annotation.Keep;
 @Keep
 public class ArchiveException extends Exception {
 
-  public ArchiveException(String message) {
-    super(message);
-  }
+    public ArchiveException(String message) {
+        super(message);
+    }
 
-  public ArchiveException(String message, Throwable cause) {
-    super(message, cause);
-  }
+    public ArchiveException(String message, Throwable cause) {
+        super(message, cause);
+    }
 }
