+++ library/src/main/java/com/hippo/a7zip/PropID.java
@@ -17,100 +17,100 @@
 package com.hippo.a7zip;
 
 public enum PropID {
-  NO_PROPERTY,
-  MAIN_SUBFILE,
-  HANDLER_ITEM_INDEX,
-  PATH,
-  NAME,
-  EXTENSION,
-  IS_DIR,
-  SIZE,
-  PACK_SIZE,
-  ATTRIB,
-  C_TIME,
-  A_TIME,
-  M_TIME,
-  SOLID,
-  COMMENTED,
-  ENCRYPTED,
-  SPLIT_BEFORE,
-  SPLIT_AFTER,
-  DICTIONARY_SIZE,
-  CRC,
-  TYPE,
-  IS_ANTI,
-  METHOD,
-  HOST_OS,
-  FILE_SYSTEM,
-  USER,
-  GROUP,
-  BLOCK,
-  COMMENT,
-  POSITION,
-  PREFIX,
-  NUM_SUB_DIRS,
-  NUM_SUB_FILES,
-  UNPACK_VER,
-  VOLUME,
-  IS_VOLUME,
-  OFFSET,
-  LINKS,
-  NUM_BLOCKS,
-  NUM_VOLUMES,
-  TIME_TYPE,
-  BIT64,
-  BIG_ENDIAN,
-  CPU,
-  PHY_SIZE,
-  HEADERS_SIZE,
-  CHECKSUM,
-  CHARACTS,
-  VA,
-  ID,
-  SHORT_NAME,
-  CREATOR_APP,
-  SECTOR_SIZE,
-  POSIX_ATTRIB,
-  SYM_LINK,
-  ERROR,
-  TOTAL_SIZE,
-  FREE_SPACE,
-  CLUSTER_SIZE,
-  VOLUME_NAME,
-  LOCAL_NAME,
-  PROVIDER,
-  NT_SECURE,
-  IS_ALT_STREAM,
-  IS_AUX,
-  IS_DELETED,
-  IS_TREE,
-  SHA1,
-  SHA256,
-  ERROR_TYPE,
-  NUM_ERRORS,
-  ERROR_FLAGS,
-  WARNING_FLAGS,
-  WARNING,
-  NUM_STREAMS,
-  NUM_ALT_STREAMS,
-  ALT_STREAMS_SIZE,
-  VIRTUAL_SIZE,
-  UNPACK_SIZE,
-  TOTAL_PHY_SIZE,
-  VOLUME_INDEX,
-  SUB_TYPE,
-  SHORT_COMMENT,
-  CODE_PAGE,
-  IS_NOT_ARC_TYPE,
-  PHY_SIZE_CANT_BE_DETECTED,
-  ZEROS_TAIL_IS_ALLOWED,
-  TAIL_SIZE,
-  EMBEDDED_STUB_SIZE,
-  NT_REPARSE,
-  HARD_LINK,
-  INODE,
-  STREAM_ID,
-  READ_ONLY,
-  OUT_NAME,
-  COPY_LINK,
+    NO_PROPERTY,
+    MAIN_SUBFILE,
+    HANDLER_ITEM_INDEX,
+    PATH,
+    NAME,
+    EXTENSION,
+    IS_DIR,
+    SIZE,
+    PACK_SIZE,
+    ATTRIB,
+    C_TIME,
+    A_TIME,
+    M_TIME,
+    SOLID,
+    COMMENTED,
+    ENCRYPTED,
+    SPLIT_BEFORE,
+    SPLIT_AFTER,
+    DICTIONARY_SIZE,
+    CRC,
+    TYPE,
+    IS_ANTI,
+    METHOD,
+    HOST_OS,
+    FILE_SYSTEM,
+    USER,
+    GROUP,
+    BLOCK,
+    COMMENT,
+    POSITION,
+    PREFIX,
+    NUM_SUB_DIRS,
+    NUM_SUB_FILES,
+    UNPACK_VER,
+    VOLUME,
+    IS_VOLUME,
+    OFFSET,
+    LINKS,
+    NUM_BLOCKS,
+    NUM_VOLUMES,
+    TIME_TYPE,
+    BIT64,
+    BIG_ENDIAN,
+    CPU,
+    PHY_SIZE,
+    HEADERS_SIZE,
+    CHECKSUM,
+    CHARACTS,
+    VA,
+    ID,
+    SHORT_NAME,
+    CREATOR_APP,
+    SECTOR_SIZE,
+    POSIX_ATTRIB,
+    SYM_LINK,
+    ERROR,
+    TOTAL_SIZE,
+    FREE_SPACE,
+    CLUSTER_SIZE,
+    VOLUME_NAME,
+    LOCAL_NAME,
+    PROVIDER,
+    NT_SECURE,
+    IS_ALT_STREAM,
+    IS_AUX,
+    IS_DELETED,
+    IS_TREE,
+    SHA1,
+    SHA256,
+    ERROR_TYPE,
+    NUM_ERRORS,
+    ERROR_FLAGS,
+    WARNING_FLAGS,
+    WARNING,
+    NUM_STREAMS,
+    NUM_ALT_STREAMS,
+    ALT_STREAMS_SIZE,
+    VIRTUAL_SIZE,
+    UNPACK_SIZE,
+    TOTAL_PHY_SIZE,
+    VOLUME_INDEX,
+    SUB_TYPE,
+    SHORT_COMMENT,
+    CODE_PAGE,
+    IS_NOT_ARC_TYPE,
+    PHY_SIZE_CANT_BE_DETECTED,
+    ZEROS_TAIL_IS_ALLOWED,
+    TAIL_SIZE,
+    EMBEDDED_STUB_SIZE,
+    NT_REPARSE,
+    HARD_LINK,
+    INODE,
+    STREAM_ID,
+    READ_ONLY,
+    OUT_NAME,
+    COPY_LINK,
 }
